// src/utils/gridViewPayLoad.d.ts
export interface gridViewPayLoadSuggestedPos{
    gridState?: {
        columns?: [
            {
                visible?: boolean,
                width?: string,
                name?: string,
                sort?:{
                    direction: string
                }
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
            {
                visible?: boolean,
                width?: string,
                name?: string,
            },
        ],
        scrollFocus?: {},
        selection?: [],
        grouping?: {},
        treeView?: {},
        pagination?:{
            paginationCurrentPage?: Number,
            paginationPageSize?: Number
        },
        filter?: {
            filters?: [],
            logic?: string
        },
        sort?: [
            {
                dir?: string,
                field?: string,
            }
        ]
    }
    name?: string
    itemType?: string 
};